+++ receiver/otlpreceiver/otlp_test.go
@@ -69,7 +69,7 @@ var traceJSON = []byte(`
 			  }
 			]
 		  },
-		  "instrumentation_library_spans": [
+		  "scope_spans": [
 			{
 			  "spans": [
 				{
@@ -365,8 +365,8 @@ func testHTTPJSONRequest(t *testing.T, url string, sink *internalconsumertest.Er
 	allTraces := sink.AllTraces()
 	if expectedErr == nil {
 		assert.Equal(t, 200, resp.StatusCode)
-		_, err = otlpgrpc.UnmarshalJSONTracesResponse(respBytes)
-		assert.NoError(t, err, "Unable to unmarshal response to TracesResponse")
+		tr := otlpgrpc.NewTracesResponse()
+		assert.NoError(t, tr.UnmarshalJSON(respBytes), "Unable to unmarshal response to TracesResponse")
 
 		require.Len(t, allTraces, 1)
 		assert.EqualValues(t, allTraces[0], traceOtlp)
@@ -486,8 +486,9 @@ func testHTTPProtobufRequest(
 
 	if expectedErr == nil {
 		require.Equal(t, 200, resp.StatusCode, "Unexpected return status")
-		_, err = otlpgrpc.UnmarshalTracesResponse(respBytes)
-		require.NoError(t, err, "Unable to unmarshal response to TracesResponse")
+
+		tr := otlpgrpc.NewTracesResponse()
+		assert.NoError(t, tr.UnmarshalProto(respBytes), "Unable to unmarshal response to TracesResponse")
 
 		require.Len(t, allTraces, 1)
 		assert.EqualValues(t, allTraces[0], wantData)